+++ README.md
@@ -3,10 +3,11 @@
 ```bash
 26/04/2021
 
-Tambah Transaksi Modal
-Tambah Table Sorting
-Ganti Sorting Dashboard
-Pindah Pengaturan & Logout
-Clean Code, No Warning & Error
-Tambah Branch "Back"
+Change Directory and Files Format
+Change Router Format
+Add Recoil
+Add UserAuth Model
+Add Middleware
+Add Loading
+Add Not Found Page
 ```
