+++ package.json
@@ -16,6 +16,7 @@
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "recharts": "^2.0.9",
+    "recoil": "^0.2.0",
     "rsuite-table": "^3.14.5",
     "web-vitals": "^1.1.1"
   },
