--- src/containers/Barang.jsx
@@ -1,33 +1,0 @@
-import React, { Component } from "react";
-import DataTable from "../components/DataTable";
-import { CgAdd } from "react-icons/cg";
-
-class Barang extends Component {
-  render() {
-    const itemData = [
-      { id: 1, name: "Nanas", category: "Buah", count: 5, price: 5000 },
-      { id: 2, name: "Semangka", category: "Buah", count: 2, price: 25000 },
-      { id: 3, name: "Tomat", category: "Sayur", count: 8, price: 15000 },
-      { id: 4, name: "Apel", category: "Buah", count: 12, price: 12000 },
-    ];
-
-    return (
-      <div className="content px-4">
-        <div className="pb-2 is-flex is-justify-content-space-between">
-          <span className="is-size-5">Barang</span>
-          <button className="button is-info is-small is-light">
-            <i className="icon">
-              <CgAdd className="react-icons-sidebar" />
-            </i>
-            <span>Tambah Data Barang</span>
-          </button>
-        </div>
-        <div className="">
-          <DataTable listdata={itemData} />
-        </div>
-      </div>
-    );
-  }
-}
-
-export default Barang;
