--- src/containers/Pengaturan.jsx
@@ -1,19 +1,0 @@
-import React, { Component } from "react";
-
-class Pengaturan extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {};
-  }
-  render() {
-    return (
-      <div className="content px-4">
-        <div className="pb-4">
-          <span className="is-size-5">Pengaturan</span>
-        </div>
-      </div>
-    );
-  }
-}
-
-export default Pengaturan;
