+++ src/index.js
@@ -1,5 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom";
+import App from './App';
+import { RecoilRoot } from 'recoil';
+import axios from 'axios';
 
 import "./styles/index.css";
 import "./styles/colors.css";
@@ -10,11 +13,20 @@ import "bulma/css/bulma.min.css";
 import "animate.css/animate.css";
 
 import reportWebVitals from "./reportWebVitals";
-import HomePage from "./layouts/HomePage";
 
-ReactDOM.render(<HomePage />, document.getElementById("root"));
+axios.defaults.baseURL = 'http://127.0.0.1:8000/api';
+axios.interceptors.request.use((config) => {
+    config.headers.Authorization = `Bearer ${localStorage.getItem('userToken')}`;
+    return config;
+});
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+ReactDOM.render(
+    <React.StrictMode>
+        <RecoilRoot>
+            <App />
+        </RecoilRoot>
+    </React.StrictMode>,
+    document.getElementById('root')
+);
+
 reportWebVitals();
