+++ src/middleware/index.jsx
@@ -1,0 +1,4 @@
+import Authenticated from './Authenticated';
+import Guest from './Guest';
+
+export { Authenticated, Guest };
